+++ lib/nnq/bus.rb
@@ -13,6 +13,7 @@ module NNQ
   #
   class BUS0 < Socket
     def send(body)
+      body = frozen_binary(body)
       Reactor.run { @engine.routing.send(body) }
     end