+++ lib/nnq/pair.rb
@@ -10,6 +10,7 @@ module NNQ
   #
   class PAIR0 < Socket
     def send(body)
+      body = frozen_binary(body)
       Reactor.run { @engine.routing.send(body) }
     end