+++ lib/nnq/push_pull.rb
@@ -11,6 +11,7 @@ module NNQ
   #
   class PUSH0 < Socket
     def send(body)
+      body = frozen_binary(body)
       Reactor.run { @engine.routing.send(body) }
     end